refactor(MP-99): Move board holder component to organisms folder.
